Sign assemblies
This change creates strong-named versions of the assemblies. While this
doesn't have any benefits in regards to security it helps those projects
that (for whatever reason) are required to create strong-named assemblies
because all dependencies have to be strong-named as well.

Since signing an assembly doesn't help in any ways in regards to security,
it doesn't hurt if we add the key file to the repo.